requirements.txt
pandas>=2.0.0
numpy>=1.24.0
kafka-python>=2.0.2
pyyaml>=6.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
eventlet>=0.33.0
TA-Lib>=0.4.28
pytest>=7.4.0
